MainActivity 수정, MypageActivity 추가
